Did some refactoring for nodes and pipeline. Trying to split apart Relion specific stuff from more generic parts. This is still ongoing.
