+++ setup.py
@@ -2,7 +2,7 @@
 
 import setuptools
 
-console_scripts = ["relipy.show=relion.cli.pipeline_viewer:run"]
+console_scripts = ["relipy.show=relion.pipeline_viewer:run"]
 setuptools.setup(
     entry_points={"console_scripts": console_scripts},
 )
