+++ src/relion/__init__.py
@@ -9,7 +9,8 @@ from relion._parser.ctffind import CTFFind
 from relion._parser.motioncorrection import MotionCorr
 from relion._parser.class2D import Class2D
 from relion._parser.class3D import Class3D
-from relion._parser.pipeline import Pipeline, RelionNode
+from relion._parser.relion_pipeline import RelionPipeline
+from relion._parser.pipeline import ProcessNode
 
 __all__ = []
 __author__ = "Diamond Light Source - Scientific Software"
@@ -33,7 +34,9 @@ class Project:
         self.basepath = pathlib.Path(path)
         if not self.basepath.is_dir():
             raise ValueError(f"path {self.basepath} is not a directory")
-        self.pipeline = Pipeline(self.basepath / "Movies", RelionNode("Import/job001"))
+        self.pipeline = RelionPipeline(
+            self.basepath / "Movies", ProcessNode("Import/job001")
+        )
 
     def __eq__(self, other):
         if isinstance(other, Project):
@@ -93,3 +96,8 @@ class Project:
     @property
     def schedule_files(self):
         return self.basepath.glob("pipeline*.log")
+
+    @property
+    def current_job(self):
+        self._load_pipeline()
+        return self.pipeline.current_job()
